requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0,<2.0  # Pin below 2.0 for DeepChecks compatibility
scikit-learn>=1.3.0
joblib>=1.3.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ML enhancements
imbalanced-learn>=0.11.0
category-encoders>=2.6.0
xgboost>=2.0.0

# API framework (for FastAPI later)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Workflow orchestration (for Prefect later)
prefect>=2.14.0

# Testing (for automated testing later)
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0
deepchecks>=0.17.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
pyyaml>=6.0
requests>=2.31.0